pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anomaly_engine"
version = "1.0.0"
description = "A real-time anomaly detection engine with LLM insights."
requires-python = ">=3.10,<3.13"

dependencies = [
    "python-dotenv",
    "pandas",
    "psycopg2-binary",
    "kafka-python",
    "faker",
    "pyspark",
    "langchain",
    "langchain-community",
    "langchain-postgres",
    "langchain-ollama",
    "ollama",
    "streamlit",
    "plotly"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["anomaly_engine*"]
exclude = ["data*", "jars*", "scripts*"]